requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-google-genai>=0.0.5
faiss-cpu>=1.7.4
google-generativeai>=0.3.1
python-dotenv>=1.0.0
arxiv>=1.4.8
PyPDF2>=3.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0